refactor: 도메인 클래스를 test에서 main으로 이동
