+++ README.md
@@ -17,7 +17,7 @@
 - 로또 게임 
   - [x] 입력된 금액에 따라 로또를 구매한다 
     - [x] 천원 단위가 아니면 예외를 발생 -> 추후 게임 진행 시 재입력을 유도한다
-    - [] 로또는 1~45 랜덤 숫자 6개로 이루어져 있어야 한다 
+    - [x] 로또는 1~45 랜덤 숫자 6개로 이루어져 있어야 한다 
   - [] 입력된 당첨 번호를 저장한다 
   - [] 구매한 로또 목록과 당첨 번호를 비교한다
     - [] 등수에 따라 나눈다
