+++ src/main/kotlin/lotto/Lotto.kt
@@ -1,0 +1,3 @@
+package lotto
+
+class Lotto(val numbers: List<Int>)
